requirements.txt
click
ipython
matplotlib
numpy
opencv-python
pysimplegui
pyyaml
scipy
scikit-image
